Move types file and ignore test files in bundle
